Merging preview to stable.

+++ src/Events/MessageEvent.ts
@@ -20,12 +20,7 @@ export async function onMessageEvent(msg: Message): Promise<void> {
 		return;
 	}
 
-	if (msg.type !== "DEFAULT") {
-		return;
-	}
-
-	// ensure no bot.
-	if (msg.author.bot) {
+	if (msg.type !== "DEFAULT" || msg.author.bot) {
 		return;
 	}
 
@@ -35,6 +30,9 @@ export async function onMessageEvent(msg: Message): Promise<void> {
 	}
 
 	if (msg.guild !== null) {
+		if (BotConfiguration.exemptGuild.includes(msg.guild.id)) {
+			return;
+		}
 		const mongoGuild: MongoDbHelper.MongoDbGuildManager = new MongoDbHelper.MongoDbGuildManager(msg.guild.id);
 		await commandHandler(msg, await mongoGuild.findOrCreateGuildDb());
 	}
@@ -113,6 +111,14 @@ async function commandHandler(msg: Message, guildHandler: IRaidGuild | null): Pr
 	if (msg.guild !== null) {
 		// because this is a guild, we have the following vars as NOT null
 		guildHandler = guildHandler as IRaidGuild;
+		if (typeof guildHandler.properties.blockedCommands !== "undefined"
+			&& guildHandler.properties.blockedCommands.includes(command.getMainCommandName()) 
+			&& !(msg.member as GuildMember).hasPermission("ADMINISTRATOR")) {
+			embed.setTitle("**Command Blocked**")
+				.setDescription("Your server administrator has blocked the use of this command. Please try again later.");
+			MessageUtil.send(embed, msg.channel, 8 * 1000).catch(() => { });
+			return;
+		}
 
 		if (command.isServerOwnerOnly() && msg.author.id !== msg.guild.ownerID) {
 			embed.setTitle("**Server Owner Command Only**")
@@ -121,15 +127,8 @@ async function commandHandler(msg: Message, guildHandler: IRaidGuild | null): Pr
 			return;
 		}
 
-		const [hasServerPerms, hasRolePerms, considerServerPerms]: [boolean, boolean, boolean] = OtherUtil.checkCommandPerms(msg, command, guildHandler);
+		let canRunCommand: boolean = OtherUtil.checkCommandPerms(msg, command, guildHandler);
 		
-		let canRunCommand: boolean;
-		if (considerServerPerms) {
-			canRunCommand = hasServerPerms || hasRolePerms;
-		}
-		else {
-			canRunCommand = hasRolePerms;
-		}
 		if (!canRunCommand) {
 			embed.setTitle("**Missing Permissions**")
 				.setDescription("You are missing either server or role permissions. Please use the help command to look up the permissions needed to run this command.")
@@ -197,7 +196,15 @@ async function commandHandler(msg: Message, guildHandler: IRaidGuild | null): Pr
 		return;
 	}
 
-	await msg.delete().catch(() => { });
+	if (msg.guild !== null && command.getDurationUntilDeleteCmd() !== -1) {
+		if (command.getDurationUntilDeleteCmd() === 0) {
+			msg.delete().catch(() => { });
+		}
+		else {
+			msg.delete({ timeout: command.getDurationUntilDeleteCmd() * 1000 }).catch(() => { });
+		}
+	}
+
 	command.executeCommand(msg, args, guildHandler);
 }
 
@@ -210,6 +217,6 @@ async function checkModMail(msg: Message): Promise<void> {
 	if (msg.guild !== null || UserAvailabilityHelper.InMenuCollection.has(msg.author.id)) {
 		return;
 	}
-	
+
 	ModMailHandler.initiateModMailContact(msg.author, msg);
 }
